pyproject.toml
[tool.poetry]
name = "kassandra-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "bot"}]

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.1.0"
aio-pika = "^8.3.0"
asyncpg = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
